feat(print-xml): 增加复杂节点与字段格式化能力
- 增加严格表格网格校验，支持表头与表体分区以及跨行跨列约束。
- 扩展 XML DSL，支持动态表格行、图片描述、书签和内部链接。
- 增加不可变格式化器注册表与 number、date、datetime 内置格式化器。
- 将格式计划固化到编译快照，并统一治理绑定错误、异常脱敏与资源上限。
- 补充复杂节点、并发复用、安全边界及 Governor 回归测试。
- 更新 XML 模块 README，明确 SpEL 与图片资源解析的后续扩展边界。

+++ letool-starter-print/src/main/java/com/github/leyland/letool/print/document/node/TableRow.java
@@ -33,8 +33,8 @@ public final class TableRow {
     }
 
     /** @return 本行所有单元格跨度之和 */
-    int effectiveColumns() {
-        return cells.stream().mapToInt(TableCell::colSpan).sum();
+    long effectiveColumns() {
+        return cells.stream().mapToLong(TableCell::colSpan).sum();
     }
 
     @Override
